+++ Hanger/.idea/misc.xml
@@ -5,12 +5,16 @@
       <map>
         <entry key="..\:/temp_project/Hanger/app/src/main/res/layout/activity_main.xml" value="0.27309782608695654" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/drawable/bg_round_rectangle.xml" value="0.1515" />
+        <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/drawable/ic_calendar.xml" value="0.1515" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/drawable/ic_login_bk.xml" value="0.1615" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/drawable/login_shape_bk.xml" value="0.1615" />
+        <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/drawable/ripple_round_rectangle.xml" value="0.1515" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/font/font_family.xml" value="0.30833333333333335" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/activity_login.xml" value="0.2" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/activity_main.xml" value="0.25" />
+        <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/activity_weather.xml" value="0.2328125" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/fragment_calendar.xml" value="0.3651041666666667" />
+        <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/fragment_cloth.xml" value="0.25" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/fragment_main.xml" value="0.25" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/fragment_weather.xml" value="0.2328125" />
         <entry key="..\:/temp_project/SE2022_3_3/Hanger/app/src/main/res/layout/layout_login.xml" value="0.18" />
