+++ Hanger/app/src/main/java/com/se/hanger/view/adapter/MainViewPagerAdapter.kt
@@ -4,10 +4,10 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.se.hanger.view.main.CalendarFragment
-import com.se.hanger.view.main.MainFragment
+import com.se.hanger.view.main.ClothFragment
 
 class MainViewPagerAdapter(fragmentActivity: FragmentActivity) : FragmentStateAdapter(fragmentActivity) {
-    private val mainFragmentList = listOf<Fragment>(MainFragment(), CalendarFragment())
+    private val mainFragmentList = listOf<Fragment>(ClothFragment(), CalendarFragment())
 
     override fun getItemCount() =mainFragmentList.size
